rename(... Waiting): registration 패키지 삭제 후 waiting, reservation 패키지를 roomescape 아래에 위치시킴
