+++ src/main/java/roomescape/admin/controller/AdminController.java
@@ -16,11 +16,11 @@ import roomescape.admin.dto.AdminReservationRequest;
 import roomescape.admin.dto.ReservationFilterRequest;
 import roomescape.auth.annotation.Auth;
 import roomescape.member.domain.MemberRole;
-import roomescape.registration.domain.reservation.dto.ReservationDto;
-import roomescape.registration.domain.reservation.dto.ReservationResponse;
-import roomescape.registration.domain.reservation.service.ReservationService;
-import roomescape.registration.domain.waiting.dto.WaitingResponse;
-import roomescape.registration.domain.waiting.service.WaitingService;
+import roomescape.reservation.dto.ReservationDto;
+import roomescape.reservation.dto.ReservationResponse;
+import roomescape.reservation.service.ReservationService;
+import roomescape.waiting.dto.WaitingResponse;
+import roomescape.waiting.service.WaitingService;
 
 @RestController
 @Auth(roles = MemberRole.ADMIN)
