+++ src/main/java/roomescape/client/payment/dto/TossPaymentConfirmRequest.java
@@ -2,7 +2,7 @@ package roomescape.client.payment.dto;
 
 import io.swagger.v3.oas.annotations.tags.Tag;
 import java.math.BigDecimal;
-import roomescape.registration.domain.reservation.dto.ReservationRequest;
+import roomescape.reservation.dto.ReservationRequest;
 
 /**
  * @see <a href="https://docs.tosspayments.com/reference#%EA%B2%B0%EC%A0%9C-%EC%8A%B9%EC%9D%B8">토스 결제 승인 API doc</a>
