+++ src/main/java/roomescape/member/controller/MemberController.java
@@ -11,10 +11,10 @@ import roomescape.member.dto.MemberResponse;
 import roomescape.member.service.MemberService;
 import roomescape.client.payment.Payment;
 import roomescape.client.payment.service.PaymentService;
-import roomescape.registration.domain.reservation.dto.ReservationDto;
-import roomescape.registration.domain.reservation.service.ReservationService;
-import roomescape.registration.domain.waiting.service.WaitingService;
-import roomescape.registration.dto.RegistrationInfoResponse;
+import roomescape.reservation.dto.ReservationDto;
+import roomescape.reservation.service.ReservationService;
+import roomescape.waiting.service.WaitingService;
+import roomescape.reservation.dto.RegistrationInfoResponse;
 
 @Tag(name = "멤버 컨트롤러", description = "모든 사용자를 반환하거나 특정 사용자의 예약 내역을 반환한다.")
 @RestController
