+++ src/main/java/roomescape/reservation/domain/Reservation.java
@@ -1,0 +1,94 @@
+package roomescape.reservation.domain;
+
+import io.swagger.v3.oas.annotations.tags.Tag;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.ManyToOne;
+import java.time.LocalDate;
+import java.util.Objects;
+import roomescape.member.domain.Member;
+import roomescape.reservationtime.domain.ReservationTime;
+import roomescape.theme.domain.Theme;
+
+@Tag(name = "예약 엔티티", description = "예약시 필요한 정보를 관리한다. 예약 생성시 특정 정보에 대해 예외처리를 진행한다.")
+@Entity
+public class Reservation {
+
+    private static final int NULL_ID = 0;
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private long id;
+
+    @Column(nullable = false)
+    private LocalDate date;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    private ReservationTime reservationTime;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    private Theme theme;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    private Member member;
+
+    public Reservation() {
+    }
+
+    public Reservation(long id, LocalDate date, ReservationTime reservationTime, Theme theme, Member member) {
+        this.id = id;
+        this.date = date;
+        this.reservationTime = reservationTime;
+        this.theme = theme;
+        this.member = member;
+    }
+
+    public Reservation(LocalDate date, ReservationTime reservationTime, Theme theme, Member member) {
+        this(NULL_ID, date, reservationTime, theme, member);
+    }
+
+    public void setMember(Member member) {
+        this.member = member;
+    }
+
+    public long getId() {
+        return id;
+    }
+
+    public LocalDate getDate() {
+        return date;
+    }
+
+    public ReservationTime getReservationTime() {
+        return reservationTime;
+    }
+
+    public Theme getTheme() {
+        return theme;
+    }
+
+    public Member getMember() {
+        return member;
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) {
+            return true;
+        }
+        if (o == null || getClass() != o.getClass()) {
+            return false;
+        }
+        Reservation that = (Reservation) o;
+        return id == that.id;
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(id);
+    }
+}
