+++ src/test/java/roomescape/reservation/controller/ReservationControllerTest.java
@@ -22,11 +22,10 @@ import roomescape.client.payment.service.PaymentClient;
 import roomescape.member.domain.Member;
 import roomescape.member.domain.MemberRole;
 import roomescape.model.ControllerTest;
-import roomescape.registration.domain.reservation.controller.ReservationController;
-import roomescape.registration.domain.reservation.domain.Reservation;
-import roomescape.registration.domain.reservation.dto.ReservationDto;
-import roomescape.registration.domain.reservation.dto.ReservationTimeAvailabilityResponse;
-import roomescape.registration.domain.reservation.service.ReservationService;
+import roomescape.reservation.domain.Reservation;
+import roomescape.reservation.dto.ReservationDto;
+import roomescape.reservation.dto.ReservationTimeAvailabilityResponse;
+import roomescape.reservation.service.ReservationService;
 import roomescape.reservationtime.domain.ReservationTime;
 import roomescape.theme.domain.Theme;
 import roomescape.vo.Name;
