+++ src/test/java/roomescape/reservation/service/ReservationServiceTest.java
@@ -24,14 +24,13 @@ import roomescape.member.domain.Member;
 import roomescape.member.domain.MemberRole;
 import roomescape.member.repository.MemberRepository;
 import roomescape.client.payment.PaymentRepository;
-import roomescape.registration.domain.reservation.domain.Reservation;
-import roomescape.registration.domain.reservation.dto.ReservationDto;
-import roomescape.registration.domain.reservation.dto.ReservationRequest;
-import roomescape.registration.domain.reservation.repository.ReservationRepository;
-import roomescape.registration.domain.reservation.service.ReservationService;
-import roomescape.registration.domain.waiting.domain.Waiting;
-import roomescape.registration.domain.waiting.repository.WaitingRepository;
-import roomescape.registration.dto.RegistrationDto;
+import roomescape.reservation.domain.Reservation;
+import roomescape.reservation.dto.ReservationDto;
+import roomescape.reservation.dto.ReservationRequest;
+import roomescape.reservation.repository.ReservationRepository;
+import roomescape.waiting.domain.Waiting;
+import roomescape.waiting.repository.WaitingRepository;
+import roomescape.reservation.dto.RegistrationDto;
 import roomescape.reservationtime.domain.ReservationTime;
 import roomescape.reservationtime.repository.ReservationTimeRepository;
 import roomescape.theme.domain.Theme;
