+++ src/test/java/roomescape/reservationtime/service/ReservationTimeServiceTest.java
@@ -23,8 +23,8 @@ import roomescape.exception.RoomEscapeException;
 import roomescape.exception.model.ReservationTimeExceptionCode;
 import roomescape.member.domain.Member;
 import roomescape.member.domain.MemberRole;
-import roomescape.registration.domain.reservation.domain.Reservation;
-import roomescape.registration.domain.reservation.repository.ReservationRepository;
+import roomescape.reservation.domain.Reservation;
+import roomescape.reservation.repository.ReservationRepository;
 import roomescape.reservationtime.domain.ReservationTime;
 import roomescape.reservationtime.dto.ReservationTimeRequest;
 import roomescape.reservationtime.dto.ReservationTimeResponse;
